feat: change slider page on marker click;
the slider is now built from a list of infos instead of using a
FutureBuilder inside it that waits for the data.
Also, notifiers are used to signal the slider to animate to a different
page, when a marker is clicked.
If the marker's info is not in the slider list, a new slider is built.